convert ints to doubles to allow for decimal point values

+++ TLBGridPoints/Data/Scripts/ShipPoints/AllGridsList.cs
@@ -24,7 +24,7 @@ namespace TLB.ShareTrack
 
         public static AllGridsList I;
 
-        public static Dictionary<string, int> PointValues = new Dictionary<string, int>();
+        public static Dictionary<string, double> PointValues = new Dictionary<string, double>();
 
 
         private static readonly Dictionary<long, IMyPlayer> AllPlayers = new Dictionary<long, IMyPlayer>();
@@ -36,7 +36,7 @@ namespace TLB.ShareTrack
 
         public static ShipTracker.NametagSettings NametagViewState = ShipTracker.NametagSettings.PlayerName;
 
-        private readonly Dictionary<string, int> _bp = new Dictionary<string, int>();
+        private readonly Dictionary<string, double> _bp = new Dictionary<string, double>();
 
         private readonly Dictionary<MyKeys, Action> _keyAndActionPairs = new Dictionary<MyKeys, Action>
         {
@@ -80,11 +80,11 @@ namespace TLB.ShareTrack
 
         private readonly List<IMyPlayer> _listPlayers = new List<IMyPlayer>();
         private readonly Dictionary<string, double> _m = new Dictionary<string, double>();
-        private readonly Dictionary<string, int> _mbp = new Dictionary<string, int>();
-        private readonly Dictionary<string, int> _mobp = new Dictionary<string, int>();
-        private readonly Dictionary<string, int> _obp = new Dictionary<string, int>();
+        private readonly Dictionary<string, double> _mbp = new Dictionary<string, double>();
+        private readonly Dictionary<string, double> _mobp = new Dictionary<string, double>();
+        private readonly Dictionary<string, double> _obp = new Dictionary<string, double>();
 
-        private readonly Dictionary<string, int> _pbp = new Dictionary<string, int>();
+        private readonly Dictionary<string, double> _pbp = new Dictionary<string, double>();
 
         // todo: remove this and replace with old solution for just combining BP and mass
         private readonly Dictionary<string, List<string>> _ts = new Dictionary<string, List<string>>();
@@ -98,10 +98,10 @@ namespace TLB.ShareTrack
         {
             try
             {
-                var dict = message as Dictionary<string, int>;
+                var dict = message as Dictionary<string, double>;
                 if (dict != null)
                 {
-                    PointValues = new Dictionary<string, int>();
+                    PointValues = new Dictionary<string, double>();
                     foreach (var kvp in dict)
                     {
                         AddFuzzyPointValue(kvp.Key, kvp.Value);
@@ -120,7 +120,7 @@ namespace TLB.ShareTrack
             }
         }
 
-        private void AddFuzzyPointValue(string key, int value)
+        private void AddFuzzyPointValue(string key, double value)
         {
             foreach (var existingKey in PointValues.Keys.ToList())
             {
@@ -129,7 +129,7 @@ namespace TLB.ShareTrack
                     PointValues[existingKey] = value;
                 }
             }
-            PointValues[key] = value; // Add the original key-value pair as well
+            PointValues[key] = value;
         }
 
         private void HudRegistered()
@@ -210,8 +210,8 @@ namespace TLB.ShareTrack
 
 
         private void MainTrackerUpdate(Dictionary<string, List<string>> ts, Dictionary<string, double> m,
-            Dictionary<string, int> bp, Dictionary<string, int> mbp, Dictionary<string, int> pbp,
-            Dictionary<string, int> obp, Dictionary<string, int> mobp)
+            Dictionary<string, double> bp, Dictionary<string, double> mbp, Dictionary<string, double> pbp,
+            Dictionary<string, double> obp, Dictionary<string, double> mobp)
         {
             foreach (var shipTracker in TrackingManager.I.TrackedGrids.Values)
             {
@@ -292,8 +292,8 @@ namespace TLB.ShareTrack
 
 
         private static void TeamBpCalc(StringBuilder tt, Dictionary<string, List<string>> trackedShip,
-            Dictionary<string, double> m, Dictionary<string, int> bp, Dictionary<string, int> mbp,
-            Dictionary<string, int> pbp, Dictionary<string, int> obp, Dictionary<string, int> mobp)
+            Dictionary<string, double> m, Dictionary<string, double> bp, Dictionary<string, double> mbp,
+            Dictionary<string, double> pbp, Dictionary<string, double> obp, Dictionary<string, double> mobp)
         {
             foreach (var faction in trackedShip.Keys)
             {
